% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect observations}
\usage{
inspect(.data, ob, cols = "all")
}
\arguments{
\item{.data}{A tbl or data frame.}

\item{ob}{integer. A row of the \code{observations} attribute to be inspected.}

\item{cols}{}
}
\value{

}
\description{
TO BE WRITTEN
}
\examples{
library(magrittr)

df <- data.frame(id = c("a", "b", "c"), x = 1:3, y = 2:4, z = c(1.1, 10, 5))
df <- df \%>\% 
  observe_if(nrow(.)==4, 
             z==floor(z), 
             y-x==1, 
             x < 3) \%>\% 
  observe_if(y < 4, 
             x > 1)
  
print(obs(df))

df1 <- df \%>\%  
  inspect(ob = 6, cols="most")
print(obs(df1))

}
\seealso{
\code{\link[observer]{observe_if}} in this package.
}

