% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{observations}
\alias{obs}
\alias{obs<-}
\alias{observations}
\alias{observations<-}
\title{Get or set the 'observations' attribute}
\usage{
observations(.data, compressed = TRUE)

observations(x) <- value

obs(.data, compressed = TRUE)

obs(x) <- value
}
\arguments{
\item{.data, x}{A tbl or data frame}

\item{compressed}{logical. By default, the column \code{"Rows"} 
of \code{observations} is compressed with the \code{\link[bit]{as.bit}} 
function from package \pkg{bit}. 
If \code{compressed=FALSE}, this column is uncompressed 
(and becomes understandable by the user).}

\item{value}{An object of class \code{obs_df}, see 
\code{\link[observer]{observations}}.}
}
\value{
A tbl.
}
\description{
The function \code{observations} (or \code{obs} for short) 
extracts the \code{observations} attribute from \code{.data} 
created by \code{\link[observer]{observe_if}} or 
\code{\link[observer]{observe_if_}}.
}
\seealso{
\code{\link[observer]{observe_if}} in this package.
}

