% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.R
\name{reobserve}
\alias{recheck}
\alias{reensure}
\alias{reobserve}
\title{Update observations}
\usage{
reobserve(.data, status = c("failed", "passed"))

recheck(.data, status = c("failed", "passed"))

reensure(.data, status = c("failed", "passed"))
}
\arguments{
\item{.data}{A tbl or data frame.}

\item{status}{character.}
}
\value{

}
\description{
blablablabla
}
\seealso{
\code{\link[observer]{observe_if}}, 
\code{\link[observer]{check_that}}, 
and \code{\link[observer]{ensure_that}} in this package.
}

