% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/View_obs.R
\name{View_obs}
\alias{View_obs}
\title{View observations}
\usage{
View_obs(x, title)
}
\arguments{
\item{x}{A data frame, possibly with an \code{observations} attribute.}

\item{title}{Title for viewer window.}
}
\description{
Invoke data viewer on \code{observations(x)}.
}
\seealso{
\code{\link[utils]{View}} from package \pkg{utils}.
}

