% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.R
\name{reobserve}
\alias{recheck}
\alias{reensure}
\alias{reobserve}
\title{Update observations}
\usage{
reobserve(.data, status = c("failed", "passed"))

recheck(.data, status = c("failed", "passed"))

reensure(.data, status = c("failed", "passed"))
}
\arguments{
\item{.data}{A tbl or data frame.}

\item{status}{character. One of \code{"failed"}, \code{"passed"}, or both. 
The \code{observations} attribute of \code{.data} is filtered according 
to the value of \code{status}.}
}
\value{
\code{.data} is returned, with its \code{observations} attribute udpated.
}
\description{
\code{reobserve} (resp. \code{recheck}, \code{reensure}) re-reruns the 
function \code{observe_if} (resp. \code{check_that}, \code{ensure_that}) on 
\code{.data} with respect to the predicates contained in its 
\code{observations} attribute created at a previous occasion.
}
\seealso{
\code{\link[observer]{observe_if}}, 
\code{\link[observer]{check_that}}, 
and \code{\link[observer]{ensure_that}} in this package.
}

